refactor: TaskCreateRequest 리팩토링 및 Schedule 필드 추가
1. 패키지 경로 수정 및 presentation 패키지로 이동한다.
2. startDate, dueDate 필드 제거하고 Schedule 필드로 통합한다.
3. 불필요한 검증 로직 제거한다.

Resolves #22
